refactor : Post Service 4개를 PostService로 통합
